# servidor/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API de Monitoreo de Estaciones e Infraestructura (ICMd)",
    "version": "1.0.0",
    "description": "Documentación interactiva de la API REST para la gestión y monitoreo en tiempo real de estaciones de trabajo, clases, telemetría y métricas de salud.\n\n### Autenticación\nLa mayoría de las rutas de esta API están protegidas mediante **JWT (JSON Web Token)**. \nPara probar los endpoints protegidos en esta interfaz:\n1. Utiliza el endpoint `POST /api/auth/login` con tus credenciales de administrador.\n2. Copia el token retornado en la respuesta.\n3. Haz clic en el botón **Authorize** (arriba a la derecha) e ingresa el token JWT.\n4. Swagger UI agregará automáticamente el encabezado `Authorization: Bearer <token>` en cada solicitud.",
    "contact": {
      "name": "Equipo de Desarrollo ICMd"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desarrollo local"
    },
    {
      "url": "/",
      "description": "Servidor actual (relativo)"
    }
  ],
...
  "paths": {
    "/api/auth/": {
      "get": {
        "tags": ["Auth"],
        "summary": "Listar usuarios registrados",
        "description": "Obtiene la lista de todos los usuarios registrados en el sistema. Endpoint público / de consulta administrativa.",
        "operationId": "obtenerUsuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno al listar usuarios.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "No fue posible listar los usuarios."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrar usuario administrador inicial",
        "description": "Registra el único usuario administrador del sistema. Si ya existe un administrador registrado, el sistema rechazará la solicitud con un código 403 Forbidden. La contraseña debe tener mínimo 8 caracteres, al menos una letra mayúscula y al menos un número.",
        "operationId": "crearUsuario",
        "requestBody": {
          "description": "Datos del administrador a registrar",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "name": "Administrador Principal",
                "email": "admin@gmail.com",
                "password": "Password123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario administrador creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Administrador Principal"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "admin@gmail.com"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-09-27T15:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos o faltantes (campos obligatorios, formato de correo o requisitos de contraseña).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "camposObligatorios": {
                    "summary": "Campos incompletos",
                    "value": {
                      "mensaje": "Todos los campos son obligatorios"
                    }
                  },
                  "contrasenaInvalida": {
                    "summary": "Contraseña no cumple requisitos",
                    "value": {
                      "mensaje": "Contraseña inválida"
                    }
                  },
                  "correoInvalido": {
                    "summary": "Formato de correo inválido",
                    "value": {
                      "mensaje": "Correo inválido"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Ya existe un administrador en el sistema.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "El sistema ya tiene un administrador registrado"
                }
              }
            }
          },
          "409": {
            "description": "El correo electrónico ya se encuentra registrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Correo ya existe"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "No fue posible registrar el usuario."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Iniciar sesión de usuario",
        "description": "Autentica las credenciales de un usuario registrado y genera un token JWT con vigencia de 1 hora para acceder a los endpoints protegidos de la API.",
        "operationId": "iniciarSesion",
        "requestBody": {
          "description": "Credenciales de acceso",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "admin@gmail.com",
                "password": "Password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso. Retorna el token JWT y la información del usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Campos faltantes en el cuerpo de la solicitud.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "Todos los campos son obligatorios"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas (correo o contraseña incorrectos).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "Credenciales inválidas"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor al procesar el inicio de sesión.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "Error interno del servidor"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/summary": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Obtener resumen y métricas consolidadas del dashboard",
        "description": "Calcula y retorna los contadores en tiempo real del estado de las estaciones pertenecientes a una clase: total de activas, estaciones en OK, advertencia, crítico e ignoradas.\n\n**Autenticación:** Requiere Bearer Token JWT.",
        "operationId": "getResumen",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "description": "ID numérico de la clase para consolidar el resumen",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Resumen métrico generado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "400": {
            "description": "Falta el parámetro requerido classId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Falta classId"
                }
              }
            }
          },
          "401": {
            "description": "Token no proporcionado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "Token no proporcionado."
                }
              }
            }
          },
          "403": {
            "description": "Token inválido o expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "mensaje": "Token inválido o expirado."
                }
              }
            }
          },
          "404": {
            "description": "No existe una clase con el ID especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "No existe una clase con ese ID"
                }
              }
            }
          },
          "500": {
            "description": "Error interno al generar el resumen del dashboard.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "No fue posible mostrar el resumen."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Introduce el token JWT obtenido tras autenticarte en `POST /api/auth/login`. No incluyas la palabra 'Bearer ', Swagger UI lo enviará automáticamente."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "description": "Entidad de usuario del sistema",
        "required": ["id", "name", "email", "passwordHash"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único incremental del usuario",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nombre completo del usuario",
            "example": "Alejandro Admin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Dirección de correo electrónico única",
            "example": "admin@gmail.com"
          },
          "passwordHash": {
            "type": "string",
            "description": "Hash bcrypt de la contraseña",
            "example": "$2b$10$EPwX1p2K9YgQ9e7wL0hN5.k.Jd5P4r6b/c8a2b1c3d4e5f6g7h8i"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha y hora de registro del usuario",
            "example": "2026-09-27T14:30:00.000Z"
          }
        }
      },
...
      "DashboardSummary": {
        "type": "object",
        "description": "Resumen métrico de estado y conteo de estaciones de una clase",
        "required": ["classId", "active", "ok", "warning", "critical", "ignored"],
        "properties": {
          "classId": {
            "type": "integer",
            "description": "ID de la clase consultada",
            "example": 1
          },
          "active": {
            "type": "integer",
            "description": "Total de estaciones activas (no ignoradas)",
            "example": 18
          },
          "ok": {
            "type": "integer",
            "description": "Estaciones en estado OK (reporte reciente <= 25s)",
            "example": 15
          },
          "warning": {
            "type": "integer",
            "description": "Estaciones en estado ADVERTENCIA (reporte entre 25s y 40s)",
            "example": 2
          },
          "critical": {
            "type": "integer",
            "description": "Estaciones en estado CRITICO (> 40s o falla declarada INTERNET/IA)",
            "example": 1
          },
          "ignored": {
            "type": "integer",
            "description": "Estaciones con bandera de ignoradas",
            "example": 2
          }
        }
      },
      "Error": {
        "type": "object",
        "description": "Estructura estándar de respuesta para errores de la API",
        "properties": {
          "error": {
            "type": "string",
            "description": "Descripción detallada del error",
            "example": "Mensaje descriptivo del error"
          },
          "mensaje": {
            "type": "string",
            "description": "Mensaje informativo del error o advertencia",
            "example": "Credenciales inválidas"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "description": "Cuerpo de solicitud para registro de administrador inicial",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre completo del administrador",
            "example": "Administrador Principal"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo electrónico válido",
            "example": "admin@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña segura (mínimo 8 caracteres, al menos 1 letra mayúscula y 1 número)",
            "example": "Password123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "Cuerpo de solicitud para inicio de sesión",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Correo del usuario",
            "example": "admin@gmail.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "Password123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "description": "Respuesta con token JWT e información del usuario autenticado",
        "required": ["token", "expiresIn", "user"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Token JWT de autenticación Bearer",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJhZG1pbkBnbWFpbC5jb20iLCJpYXQiOjE3OTA0ODk0NjQsImV4cCI6MTc5MDQ5MzA2NH0.tuodhgjOm-_H97LfN2KIWJ2ct4-ukPGSkPQopDk4e9U"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Tiempo de expiración del token en segundos (3600 = 1 hora)",
            "example": 3600
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "name": {
                "type": "string",
                "example": "Administrador"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "admin@gmail.com"
              }
            }
          }
        }
      },
...
    }
  }
}